fixed issue with db not connecting. Again.
