+++ app/hindsite/tables/card.py
@@ -5,7 +5,7 @@ Defines the model for the cards and their relationships to others tables in the 
 from typing import Optional
 from sqlalchemy import ForeignKey, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from app.hindsite.extensions import intpk, db
+from app.hindsite.tables.base import db, intpk
 
 
 class Card(db.Model):  # pylint: disable=too-few-public-methods
@@ -18,7 +18,7 @@ class Card(db.Model):  # pylint: disable=too-few-public-methods
     field_id: Mapped[int] = mapped_column(ForeignKey('field.id'))
     author_id: Mapped[int] = mapped_column(ForeignKey('user.id'))
     owner_id: Mapped[Optional[int]] = mapped_column(ForeignKey('user.id'))
-    message_body: Mapped[str] = mapped_column(String(64000))
+    message_body: Mapped[str] = mapped_column(String(6000))
     card_status: Mapped[str] = mapped_column(String(50))
     archived: Mapped[bool] = mapped_column(default=False)
     field = relationship('Field', back_populates='cards')
