+++ app/hindsite/tables/user.py
@@ -6,7 +6,7 @@ import datetime
 from typing import Optional
 from sqlalchemy import DateTime, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from app.hindsite.extensions import db, intpk, user_membership
+from app.hindsite.tables.base import db, intpk, user_membership
 
 
 class User(db.Model):  # pylint: disable=too-few-public-methods
